pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iplotx"
description = "Plot networkx from igraph and networkx."
authors = [
  {name = "Fabio Zanini", email = "fabio.zanini@unsw.edu.au"}
]
maintainers = [
  {name = "Fabio Zanini", email = "fabio.zanini@unsw.edu.au"}
]
requires-python = ">=3.11"
readme = "README.md"
dynamic = ["version"]
dependencies = [
  "matplotlib>=2.0.0",
  "numpy>=2.0.0",
  "pandas>=2.0.0",
  "pylint>=3.3.7",
]
license = {text = "MIT"}
license-files = ["LICEN[CS]E.*"]
keywords = ["network", "graph", "visualisation", "plotting"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Intended Audience :: Developers",
  "Intended Audience :: Education",
  "Intended Audience :: Science/Research",
  "Natural Language :: English",
  "Topic :: Scientific/Engineering :: Visualization",
  "Topic :: System :: Networking",
  "Typing :: Typed",
  ]

[project.optional-dependencies]
networkx = [
  "networkx>=2.0.0",
]
igraph = [
  "igraph>=0.11.0",
]

[dependency-groups]
test = [
  "pytest>=7.0.0",
  "flake8>=3.9.0",
  "igraph>=0.11.8",
  "networkx>=3.4.2",
  "biopython>=1.85",
  "cogent3>=2025.5.8a6",
  "ete4>=4.3.0",
  "mypy>=1.15.0",
  "scikit-bio>=0.6.3",
  "pandas-stubs>=2.2.3.250527",
  "types-networkx>=3.5.0.20250610",
]
docs = [
  "sphinx",
  "sphinx-rtd-theme>=3.0.2",
  "myst-parser>=4.0.1",
  "sphinx-design>=0.6.1",
  "sphinx-gallery>=0.19.0",
  "igraph>=0.11.8",
  "networkx>=3.4.2",
  "biopython>=1.85",
  "cogent3>=2025.5.8a6",
  "ete4>=4.3.0",
  "scikit-bio>=0.6.3",
]
typing = ["mypy", "types-requests"]
typing-test = [{include-group = "typing"}, {include-group = "test"}, "useful-types"]

[project.urls]
Homepage = "https://github.com/fabilab/iplotx"
Documentation = "https://readthedocs.org/iplotx"
Repository = "https://github.com/fabilab/iplotx.git"
"Bug Tracker" = "https://github.com/fabilab/iplotx/issues"
Changelog = "https://github.com/fabilab/iplotx/blob/main/CHANGELOG.md"

[tool.hatch.version]
path = "iplotx/version.py"

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[tool.ruff]
line-length = 100
extend-exclude = ["gallery/**/plot_*.py"]
